Move existing RTK tests into new `./tests` subfolders
