+++ src/tests/combinedTest.test.ts
@@ -1,10 +1,10 @@
-import { createAsyncThunk } from './createAsyncThunk'
-import { createAction, PayloadAction } from './createAction'
-import { createSlice } from './createSlice'
-import { configureStore } from './configureStore'
-import { createEntityAdapter } from './entities/create_adapter'
-import { EntityAdapter } from './entities/models'
-import { BookModel } from './entities/fixtures/book'
+import { createAsyncThunk } from '../createAsyncThunk'
+import { createAction, PayloadAction } from '../createAction'
+import { createSlice } from '../createSlice'
+import { configureStore } from '../configureStore'
+import { createEntityAdapter } from '../entities/create_adapter'
+import { EntityAdapter } from '../entities/models'
+import { BookModel } from '../entities/tests/fixtures/book'
 
 describe('Combined entity slice', () => {
   let adapter: EntityAdapter<BookModel>
